1.2.3: EXPERIMENTAL build for windows-x86_64
* tested only with windows 10 & 7 on Virtual Machines.
  + In this Windows release the HOME dir (where user data including
    savegames are stored) is now in Windows UserProfile Folder instead of the data
    directory (eg: 'c:\users\NAME\Local Settings\privgold120').
  + Currently non-utf8 characters written to console are truncated and cause
    exception in python (which is caught), when writing on the given file (stdout/stderr).
  + the std libc chdir()/fopen()/... do not handle utf8, then i needed dirty fix to use WCHAR
    (\_wmkdir/.....) and use VSFileSystem::vs_{chdir,fopen,...} wrappers in windows version.
    That would be so good to have a POSIX windows (we can dream).
  + updates in vegastrike sources for windows/msys2 sdk and clang 13
    (system types, pointers printf formats, aldrv,
  + use SHPathMatchSpec() on windows instead of fnmatch. This is not exactly fnmatch,
    but close (misses characters set in brackets)
    Add (not used yet) code to convert a fnmatch pattern in an extended regex pattern.

* fix: winsys/SDL1 mouse workaround on some windows and linux systems (mouse was not moving)

* fix: gui/FontHeight management in cockpit when Cockpit font is not base font.

* fix: (1.03) bug: don't crash if the savegame cannot be saved for some reasons
  (eg:System does not allow name, access denied, ...)

+++ data/bases/computer_lib.py
@@ -93,10 +93,11 @@ class NewSaveGame: pass
 
 def file_exists(filename):
 	try: 
-		st = os.stat(filename)		
+		f = open(filename)
+		f.close()
+		return True		
 	except:
 		return False
-	return st is not None
 
 def time_sorted_listdir(dir):
 	import os
@@ -452,13 +453,20 @@ class QuineComputer:
 					trace(TRACE_INFO, "::: picker_screen NOT visible")
 
 					savename = self.picker_screen.items[self.picker_screen.selection].data
-					VS.saveGame(savename)
-					self.oldSaveName = savename
-
-					# redisplay saved list
-					self.picker_screen.items = [GUI.GUISimpleListPicker.listitem("New Game",NewSaveGame)]+savelist()
-					self.picker_screen.show()
-					self.txt_screen.hide()
+					if not VS.saveGame(savename):
+						self.txt_screen.setText("\n"*7
+							+ "An unknown error occured while saving game '"+savename+"'.\n"
+							+ "\n"*3
+							+ "Press SAVE to try another name.")
+						self.txt_screen.show()
+						button_index = ''
+					else:	
+						self.oldSaveName = savename
+	
+						# redisplay saved list
+						self.picker_screen.items = [GUI.GUISimpleListPicker.listitem("New Game",NewSaveGame)]+savelist()
+						self.picker_screen.show()
+						self.txt_screen.hide()
 					
 					# or should this call reset? or show "game saved" text?
 					VS.restoreKeyRepeat()
@@ -480,7 +488,7 @@ class QuineComputer:
 			#if saveitem in savelist() and ... # not working
 			if savename.lower() == "new_game" or savename.lower() == "autosave":
 				self.txt_screen.setText("\n"*7
-						+ "Please use another name for your samegame.\n"
+						+ "Please use another name for your savegame.\n"
 						+ "\n"*3
 						+ "Then press ENTER")
 				return
@@ -496,7 +504,12 @@ class QuineComputer:
 					return
 				else:
 					self.lastEnteredSavegameName = savename
-			VS.saveGame(savename)
+			if not VS.saveGame(savename):
+				self.txt_screen.setText("\n"*7
+						+ "Please use another name for your savegame.\n"
+						+ "\n"*3
+						+ "Then press ENTER")
+				return
 			if self.oldSaveName is not None and savename is not self.oldSaveName:
 				try:
 					os.remove(VS.getSaveDir() + os.sep + self.oldSaveName)
